[react-0.0.1]
Updates:
The ES6 branch is a bit stale, but will be updated soon.
The README.md file has been updated a little bit.

Pulled out my React version of webwindow, and I'm going to make
that the primary focus of this repo development.

The Vanilla version works on front-end with no server.